requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
atomic-agents==1.0.19
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
docstring_parser==0.16
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
instructor==1.7.2
Jinja2==3.1.5
jiter==0.8.2
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
multidict==6.1.0
openai==1.61.1
platformdirs==4.3.6
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
pyfiglet==1.0.2
Pygments==2.19.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
tenacity==9.0.0
textual==0.89.1
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
uc-micro-py==1.0.3
urllib3==2.3.0
yarl==1.18.3
